requirements.txt
# Custom SQL Agent Dependencies
openai
fastapi
uvicorn
pydantic
python-dotenv
